Use bulk inserts.
The subsetter spends a significant chunk of overall running time on
single-row inserts into the destination tables. This patch accumulates
rows in the `pending` attribute on each table, then periodically flushes
updates to the database when the number of accumulated rows exceeds the
`buffer` argument. Depending on the value of `buffer`, this change
reduced running time by 20-30% on a sample database.